Feature: #33: Checkout functional and redirects upon submit
